implementando regras de validacao

+++ src/DevIO.Business/Services/FornecedorService.cs
@@ -6,7 +6,9 @@ namespace DevIO.Business.Services
 {
     public class FornecedorService : BaseService, IFornecedorService
     {
-
+        private const string fornecedorInexistente = "Fornecedor não existe!";
+        private const string fornecedorComProdutosCadastrados = "O fornecedor possui produtos cadastrados!";
+        private const string fornecedorExistente = "Já existe um fornecedor com este documento informado.";
         private readonly IFornecedorRepository _fornecedorRepository;
 
         public FornecedorService(IFornecedorRepository fornecedorRepository)
@@ -18,6 +20,12 @@ namespace DevIO.Business.Services
             if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)
                 || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
 
+            if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+            {
+                Notificar(fornecedorExistente);
+                return;
+            }
+
             await _fornecedorRepository.Adicionar(fornecedor);
         }
 
@@ -25,12 +33,39 @@ namespace DevIO.Business.Services
         {
             if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return;
 
+            if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+            {
+                Notificar(fornecedorExistente);
+                return;
+            }
+
             await _fornecedorRepository.Atualizar(fornecedor);
         }
 
 
         public async Task Remover(Guid id)
         {
+            var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+            if (fornecedor == null)
+            {
+                Notificar(fornecedorInexistente);
+                return;
+            }
+
+            if (fornecedor.Produtos.Any()) 
+            {
+                Notificar(fornecedorComProdutosCadastrados);
+                return;
+            }
+
+            var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+            if (endereco != null)
+            {
+                await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+            }
+
             await _fornecedorRepository.Remover(id);
         }
